git: Use cwd for ssh key
I'm trying to run homu in a container where it's installed (as root)
to `/usr`, but we run as non-root.  This follows general best practice
that apps shouldn't be able to mutate their code.

However, this bit broke as it tried to write to `/usr`.  Let's assume
we can write to our process' cwd instead.