Add a `linear_fake_merge` option (defaulting to True)
One of my projects (projectatomic/rpm-ostree) has a protected master
branch.  I was rather horrified to realize that homu force-pushes to
master by default.  Admittedly, I understand this is the only way to
make github think the PR was merged (as opposed to just closed).

But in my mind, that's a deficiency of github - it should have an API
to allow us to "close merged with this commit".

Anyways, add an option so that those who prefer closed over force
pushing can do that today.

We at least retitle the commit so that it's more clear it was merged.