Fix travis exemption code again for status-only context
It seems that
https://github.com/servo/homu/pull/13/commits/7fa396555de01b8dd19bd78a21c64c647926e199
regressed things when using status only.

The format seems to have changed at one point to a dictionary, but
this code was treating it as a list (using `[0]`).

Since we were already iterating over the statuses, move the special
casing for "status-only-travis" up to there, and slay the massive
conditional.

Conflicts:
	homu/main.py