server: Default host to '*'
I'm running Homu in an OpenShift v3/Kubernetes cluster, and for some reason
I didn't fully debug, the default configuration for bottle with host=''
appears to try to look up the hostname to determine the bind interface.

I think this works with "host networking" typically, but it won't
in a lot of containerized setups.

What we really want to do by default is use `*` which binds to all interfaces.